requirements.txt
crapy==1.0.5
scrapyd==1.1.0
shub==2.0.2

